+++ cpf/input_types/DioptasFunctions.py
@@ -826,7 +826,7 @@ class DioptasDetector:
         all_colours = cm.seismic(np.arange(n_entries))
 
         if np.abs(maximum_value) > np.abs(minimum_value):
-            n_cut = np.int(
+            n_cut = np.int_(
                 (
                     (2 * maximum_value - (maximum_value - np.abs(minimum_value)))
                     / (2 * maximum_value)
@@ -836,7 +836,7 @@ class DioptasDetector:
             keep = n_entries - n_cut
             all_colours = all_colours[keep:]
         else:#if np.abs(maximum_value) < np.abs(minimum_value):
-            keep= np.int(
+            keep= np.int_(
                 (
                     (
                         2 * np.abs(minimum_value)
